+++ packages/graphql-tool-utilities/src/augmentations.ts
@@ -1,65 +1,50 @@
-import {existsSync} from 'fs';
 import {GraphQLProjectConfig} from 'graphql-config/lib/GraphQLProjectConfig';
+import {
+  resolvePathRelativeToConfig,
+  resolveProjectName,
+  resolveSchemaPath,
+} from './utilities';
 
 declare module 'graphql-config/lib/GraphQLProjectConfig' {
+  /* eslint-disable lines-around-comment */
   interface GraphQLProjectConfig {
+    /**
+     * @deprecated use the utility function by the same name instead
+     */
     resolvePathRelativeToConfig(relativePath: string): string;
+    /**
+     * @deprecated use the utility function by the same name instead
+     */
     resolveProjectName(defaultName?: string): string;
+    /**
+     * @deprecated use the utility function by the same name instead
+     */
     resolveSchemaPath(ignoreMissing?: boolean): string;
   }
+  /* eslint-enable lines-around-comment */
 }
 
-export const defaultGraphQLProjectName = 'GraphQL';
-
-function resolveProjectName(
+function resolvePathRelativeToConfigAugmentation(
   this: GraphQLProjectConfig,
-  defaultName = defaultGraphQLProjectName,
+  relativePath: string,
 ) {
-  return this.projectName || defaultName;
+  return resolvePathRelativeToConfig(this, relativePath);
 }
 
-function resolveSchemaPath(this: GraphQLProjectConfig, ignoreMissing = false) {
-  // schemaPath is nullable in graphq-config even though it cannot actually be
-  // omitted. This function simplifies access ot the schemaPath without
-  // requiring a type guard.
-  if (!this.schemaPath) {
-    // this case should never happen with a properly formatted config file.
-    // graphql-config currently does not perform any validation so it's possible
-    // for a mal-formed schema to be loaded at runtime.
-    throw new Error(
-      `Missing GraphQL schemaPath for project '${this.resolveProjectName()}'`,
-    );
-  }
-
-  // resolve fully qualified schemaPath
-  const schemaPath = this.resolveConfigPath(this.schemaPath);
-
-  if (ignoreMissing) {
-    return schemaPath;
-  }
-
-  if (!existsSync(schemaPath)) {
-    const forProject = this.projectName
-      ? ` for project '${this.projectName}'`
-      : '';
-    throw new Error(
-      [
-        `Schema not found${forProject}.`,
-        `Expected to find the schema at '${schemaPath}' but the path does not exist.`,
-        `Check '${
-          this.configPath
-        }' and verify that schemaPath is configured correctly${forProject}.`,
-      ].join(' '),
-    );
-  }
-
-  return schemaPath;
+function resolveProjectNameAugmentation(
+  this: GraphQLProjectConfig,
+  defaultName?: string,
+) {
+  return resolveProjectName(this, defaultName);
 }
 
-// temporary augmentation until `graphql-config` supports this new function
-// see: https://github.com/prisma/graphql-config/pull/113
-GraphQLProjectConfig.prototype.resolvePathRelativeToConfig =
-  GraphQLProjectConfig.prototype.resolveConfigPath;
+function resolveSchemaPathAugmentation(
+  this: GraphQLProjectConfig,
+  ignoreMissing?: boolean,
+) {
+  return resolveSchemaPath(this, ignoreMissing);
+}
 
-GraphQLProjectConfig.prototype.resolveProjectName = resolveProjectName;
-GraphQLProjectConfig.prototype.resolveSchemaPath = resolveSchemaPath;
+GraphQLProjectConfig.prototype.resolvePathRelativeToConfig = resolvePathRelativeToConfigAugmentation;
+GraphQLProjectConfig.prototype.resolveProjectName = resolveProjectNameAugmentation;
+GraphQLProjectConfig.prototype.resolveSchemaPath = resolveSchemaPathAugmentation;